removed components into separate folder
