+++ src/index.tsx
@@ -1,6 +1,6 @@
 export { default as useInterval } from './hooks/useInterval';
 export { default as useLatestRef } from './hooks/useLatestRef';
 export { default as useTimeout } from './hooks/useTimeout';
-export { default as Button } from './Button';
-export { Variant as ButtonVariant } from './Button';
-export * from './Button/constants';
+export { default as Button } from './components/Button';
+export { Variant as ButtonVariant } from './components/Button';
+export * from './components/Button/constants';
