+++ stories/Content.stories.tsx
@@ -1,23 +1,21 @@
 import { Meta } from '@storybook/react';
-import Content from '../src/Content';
-import Skeleton from '../src/Skeleton';
-import Stack from '../src/Stack';
+import Content from '../src/components/Content';
+import Skeleton from '../src/components/Skeleton';
+import Stack from '../src/components/Stack';
 
-export const ContentStory = () =>
+export const ContentStory = () => (
     <Content>
-        <Skeleton className='font-white'>
-            Content
-        </Skeleton>
-    </Content>;
+        <Skeleton className="font-white">Content</Skeleton>
+    </Content>
+);
 
 ContentStory.storyName = 'Default';
 
-export const ContentAsSectionStory = () =>
-    <Content as='section'>
-        <Skeleton className='font-white'>
-            Wrapper rendered as section
-        </Skeleton>
-    </Content>;
+export const ContentAsSectionStory = () => (
+    <Content as="section">
+        <Skeleton className="font-white">Wrapper rendered as section</Skeleton>
+    </Content>
+);
 
 ContentAsSectionStory.storyName = 'Use custom element';
