+++ stories/Skeleton.stories.tsx
@@ -1,14 +1,14 @@
-import {Meta} from '@storybook/react';
-import Skeleton from '../src/Skeleton';
-import '../src/Skeleton/Skeleton.css';
+import { Meta } from '@storybook/react';
+import Skeleton from '../src/components/Skeleton';
+import '../src/components/Skeleton/Skeleton.css';
 
-export const SkeletonStory = () => <Skeleton className='font-white'>Skeleton</Skeleton>;
+export const SkeletonStory = () => <Skeleton className="font-white">Skeleton</Skeleton>;
 
-SkeletonStory.storyName  = 'Skeleton';
+SkeletonStory.storyName = 'Skeleton';
 
 const meta: Meta = {
-  title: 'Layout/Skeleton',
-  component: Skeleton,
+    title: 'Layout/Skeleton',
+    component: Skeleton
 };
 
 export default meta;
